+++ packages/terra-table/src/subcomponents/_CheckMarkCell.jsx
@@ -102,9 +102,13 @@ const CheckMarkCell = ({
   const attrSpread = { 'aria-selected': isSelected };
   const attrCheck = {};
   if (isHidden || isIcon) {
-    // A user of a screenreader still need a keyboard accessible method of selection, so providing -1 index and kydown.
-    attrSpread.onKeyDown = wrappedOnKeyDownForItem(onKeyDown, onSelect, metaData);
-    attrSpread.tabIndex = '-1';
+    if (isDisabled) {
+      attrCheck['aria-disabled'] = true;
+    } else {
+      // A user of a screenreader still need a keyboard accessible method of selection, so providing -1 index and kydown.
+      attrSpread.onKeyDown = wrappedOnKeyDownForItem(onKeyDown, onSelect, metaData);
+      attrSpread.tabIndex = '-1';
+    }
   } else if (isSelectable) {
     if (isDisabled) {
       attrCheck['aria-disabled'] = true;